requirements.txt
django
django-extensions
django-recaptcha
social-auth-app-django
django-two-factor-auth
requests
psycopg2-binary
django-otp
django-otp-yubikey
drf-yasg
djangorestframework